services/integration_management/requirements_ayo.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
sqlmodel
pydantic
python-jose[cryptography]
pytest
psycopg2-binary
